requirements.txt
streamlit
pytesseract
Pillow
nltk
scikit-learn
joblib
